+++ pages/work/quotes/[...slug].js
@@ -1,10 +1,10 @@
 import { useEffect, useState } from "react";
 import { useRouter } from "next/router";
-import Header from "../../components/global-components/Header";
-import Footer from "../../components/global-components/Footer";
-import Loader from "../../components/global-components/Loader";
-import ErrorMessage from "../../components/global-components/ErrorMessage";
-import QuotesResultData from "../../components/quotespage-components/QuotesResultData";
+import Header from "../../../components/global-components/Header";
+import Footer from "../../../components/global-components/Footer";
+import Loader from "../../../components/global-components/Loader";
+import ErrorMessage from "../../../components/global-components/ErrorMessage";
+import QuotesResultData from "../../../components/quotespage-components/QuotesResultData";
 
 const Slug = () => {
   const router = useRouter();
@@ -14,13 +14,14 @@ const Slug = () => {
 
   useEffect(() => {
     const fetchData = async () => {
-      const res = await fetch(`/api/quotes/home`, {
+      const res = await fetch(`/api/work-scraper`, {
         method: "POST",
         headers: {
           "content-type": "application/json",
         },
         body: JSON.stringify({
-          queryURL: `https://www.goodreads.com/quotes/${slug}`,
+          // This URL is redirected by Goodreads
+          queryURL: `http://www.goodreads.com/work/quotes/${slug}`,
         }),
       });
       if (res.ok) {
@@ -43,7 +44,7 @@ const Slug = () => {
         {error && (
           <ErrorMessage
             status="500"
-            url={`https://www.goodreads.com/quotes/${slug}`}
+            url={`https://www.goodreads.com/work/quotes/${slug}`}
           />
         )}
         {!error && (
@@ -52,13 +53,14 @@ const Slug = () => {
             {scrapedData.error && (
               <ErrorMessage
                 status="404"
-                url={`https://www.goodreads.com/quotes/${slug}`}
+                url={`https://www.goodreads.com/work/quotes/${slug}`}
               />
             )}
-            {scrapedData.quotes && scrapedData.quotes.length === 0 && (
+
+            {scrapedData.quotes && scrapedData.name.length === 0 && (
               <ErrorMessage
                 status="ScraperError"
-                url={`https://www.goodreads.com/quotes/${slug}`}
+                url={`https://www.goodreads.com/work/quotes/${slug}`}
               />
             )}
             {scrapedData && <QuotesResultData scrapedData={scrapedData} />}